Typography of primary content
Sets the root and body font-size both to 16px to ensure that rems and
ems are based on the same base size. I used 16 because it's the browser
default, so what would be expected from another developer.

Used adjacenty sibling selectors to control the space between text. This
makes the margins less broad so theres not as many cases they'll need
to be overridden.

TODO:
- Change all em values to rem.