Set all em units to rem.
Using rem because it's a more consistent unit. While em is based on the
font-size of whatever parent element, rem is confined to the font-size
of the root (html). This ensures that these values won't change if font-
size is set in different parts of document.

Only use ems when the value actaully depends on the parents font-size.

Purposely did not touch sidenav because another dev is working on it
right now.