Adds the hex bg to the footer component
I had to make this a separate absolute element so that it could be
relatively positioned compared to the centered text. This allows it
to stick to the left side of the text at different browser window sizes
rather then sticking to some point at the edge of the screen.