All icons within content
This adds all the icons within the content.

Separated the hex icons, with the hex colored backgrounds and the
free icons that are styled red.

In implemenation, the free icons were inlined as svg so their color
could be changed from the CSS. All icons will probably have to be
inlined as svg due to this. Eventually, it'd help if there was a build
process to inline the svg rather the do this.

TODO:
- probably should inline svg all icons to stay consistent.