Separators under headings with hexagon
These were built so they could conform to many different widths
and the middle piece could be switched out. With this design, the
width of the middle piece can be changed and the lines will conform.
The hexagon class uses ems for this reason, so you can set the font
size to control it's size. This had to be done because borders can't
be set with percentages so ems are used instead. And then the size
of the hexagon is set with the parent font-size.

It can also change colors from primary to alt or any other.