Background section chevrons
These are the v shaped design of the sections.

This was done by putting svg shapes on the front and back
of the sections and specifying classes for what type
of shape at what point.

This was done with svg because doing CSS shapes at 100% width
like this seemed difficult. This was all done in the section
classes because these shapes are really part of the sections,
or the sections are these shapes.

TODO:
- Investigate way to keep svg to 100% width but also maintain
aspect ration to ensure v shape doesn't change.