Fixed CSS file structure and build
This takes the CSS in the src dir and compiles it to css in the
css dir. The css dir is not checked in but will be in the npm module.
Also included a minified css file, as this allows consumers of the lib
to use the CSS correctly without their own build process.