Normal image (png) spriting
Uses a postcss plugin to create a sprite of images and update the style
sheet to use the spritesheet. The purpose is to save on bandwidth when
the browser downloads the files.

Also only copying png files from the style src img dir. SVGs will be
handled differently after this.

TODO: it'd be nice to remove the original images from the sprites from
the img/ folder as they're no longer required.

Still needs testing