Update with new uswds, javascript
Decided with team to put google analytics in here as all cloud.gov sites
will need it and be using this. It's unlikely other orgs will use
cg-style, so it's probably OK to have here.

Javascript compliled with browserify, chosen for it's simplicity and
more complex building not required.

Tracks 18F/uwds#952-override_variables, which was updated with