Add gem updating
This configuration allows the updating of the cloudgov style gem. It
does this by cloning the gem repo into ./gem, building the assets, and
copying the assets over to the correct gem folders. From here, the
changes in the gem repo can be committed and pushed to be officially
updated.