Allows base code to be imported separately
This allows a project to do this:

```
@import 'cloudgov-style/src/css/base';
```

Now only the base code is imported, meaning the consumer project
could deal with the various components when it needs it and in
whatever way it sees fit. This will allow the deck code to use
all the base styles and import the component styles into their
own component modules dynamically.