Setup vars for correct overridding
For the web design standards, I want to keep the path assets to
what they have: `../img` and `../fonts` (notice fonts is plural). But
when a consumer of cg-style wants to reset these paths to all be
the fonts and images folder within cg-style, they will override
them to point there. At this point, cg-style and uswds assets will
be pointed at what that library sets, like `./css` and `./font`.

This library should techincally provide a way for a consumer to
ingest it with webpack without having to know that it has to reset
these variables in order for things to work (after all, the paths its
setting the variables to is within the cg-style project). TODO this
should be a future improvement.

Tracks #da05c6216b69440a88aab441eaa49890ec63b2da in
18F/web_design_standards