Add font awesome as a depedency
In the process of documenting how the sidenav element works I noticed
we didn’t have font-awesome included which resulted in the glyphs not
being displayed properly. Added the library as a dependency and
included some build scripts that are run automatically on `npm run
build`

The `copy-fontawesome-css` script is weird looking because I alter the
paths required in the CSS file to match our current file structure. We
currently use `/font` so that is where the browser should look for font
awesome fonts.

I raised some related concerns in 18F/cg-style-gem#5, which could
change some parts of this implementation