Include release script and documentation update
Releasing was kind of confusing and was many
steps so I wrote a small release script to help
us out. What the release script does is documented
in the RELEASE.md guidelines.

Changed the way we clone the gem repo so that the
script doesn't prompt the user for github creds.