Add cmd to clone gem over HTTPS and SSH
The Travis build is failing because it cannot clone the gem repo over
SSH if it doesn’t have proper keys set up (even if the repo is public).
However, the release script should use the SSH URL so that the user
doesn't need to enter their Github credentials.

I refactored the gem clone commands so that there are now two options:
`gem-clone-https` and `gem-clone-ssh` which pretty much do as you
expect. The `gem` command uses the HTTPS version.

The release script still clones over SSH.