Adds css linting with stylelint to testing
Adds linting to the test command and ensures travis runs the tests.
This is an attempt to keep our CSS codebase cleaner and closer to
the 18f guide.

Important: this change requires this PR to be merged in @18f/stylelint
and a new release to be pushed:
https://github.com/18F/stylelint-rules/pull/16. It would be best to
merge this after that all happens and ensure to update the version
number here.

There are some areas I decided to ignore the linting for now, should
be fixed in the future.

ref https://github.com/18F/cg-deck/issues/183