Solved using gulp-sass and gulp-insert
Result of this: every build components .css file has a url path of
`../../img` but when the components are built as a part of the whole
library into cloudgov-style.css, they point to `../img`.

The normal node-sass node CLI makes it very hard to compile a whole
dir of scss files (which is messed up because the CLI allows it) so I
used gulp-sass which makes it easy to src a glob. I also used another
gulp plugin to insert the variable for the path into the scss files
before they are built. The reasons for doing this are discussed here:
https://github.com/dlmanning/gulp-sass/issues/302.