Bump to v0.6.1, set USE_LOCAL_GEM
Uses an env variable USE_LOCAL_GEM for
travis and bundler.

Sometimes we want to build from the published
ruby gem and sometimes we want to build from
gem in the local directory.

For example, on production and when somebody
just pulls down the style guide for local
documentation it should just work and use the
cloudgov-style gem published to the registry.

But other times we want to use the gem that
is in the local gem directory. For example,
when an automated visual regression tool is
run against a branch, we want the gem to be
composed of assets from that branch.

When the USE_LOCAL_GEM environment variable
is set it uses the local directory, and
otherwise uses the published gem.

Also, update Gemfile.lock