Move to centered content blocking loader
After a design meeting, we decided to move towards a modal loader
that appears in center of the page and hides all unloaded content.

This is done by using a 100%, absolutely positioned div with a white
background. It's done this way to ensure no data shows up behind
the div. One concern is whether this coudl be broken when placed
in relative position elements. May have to make more modifiers for
different use cases.