Fixed spacing and color issues with activity log
The pre formatted text was sometimes on the same line as the log
text and sometimes not, creating weird formatting issues. Fixed this
by giving it a min width so it's always on different line.

Ensured log text was shown to be clickable

The code text failed 508 as there wasn't enough contrast so used the
lightest gray for the background.