Fix loading block with new grid layout
The structure of the containers on the dashboard has changed so
the layout of the loading box needs to change to accomodate that.

So before, there was a container of the correct width that was position
relative. Now based on the way the containers work, they're not the
correct width, so this should just be relative.