Dealing with complex, nested panel rows.
The overview page UI has rows within rows within rows, so added new
rules to accomodate this. It was easiest to use the > selector for all
immediately nested roles and then use one more of these selectors for
the 3rd or last nested row in this case. If theres a use case in the
future for a 4th nested row (probably shouldn't be) this code can
change then.

Also added TODOs to a bunch of the rules under panel-row, hoping they
can be moved elsewhere to reduce complexity of that component.