Refactor panel row typography
And moved it to base typography where it makes more sense. To do this,
I had to create new sans classes for each of the sans type scale
sizes as certain of these sizes didn't exist in any place. These can
be used to scale type outside of the h tags.