Change icon size and layout
So it aligns in the middle of what the design calls for in terms
of size and margin around. I changed the margin around it to ems
and less of them so when it's within text it works. I changed the
size to not have a min, which doesn't work at really small sizes.