Working to make panels consistently work
In order to reuse more components on the dashboard this work ensures
panels can be used in more ways together. Changes the rules so the
triple nested panel rows on the overview page rely more on the boxed
panel row rather then the triple nesting. This ensures the other panels
don't require this extensive nesting to work.

Adds another panel header for the row, which is different the other
header as it's much smaller. It's the text for the panels that are
tables in the UI.

Modified how margins and padding is applied to last and other panel
rows to ensure consistency. Also added top margin to panel-actions
to make it easier to place there.