Panel refactor start
I recreated the simple parts of a panel. The panel has changed in
that it won't include the data part, the rows of information. It
will just include header, title actions and documentation.

I moved the feature of the panel, which will show a full example of the
panel with the rows to a separate name of panel full. I had to do this
because it seemed when you have folder/file of the same name in fractal
it gets them confused and you get the same result of both. I think they
should be separated to show how to use the panel in isolation from the
data rows which may be used alone.

I moved documentation into it's own component because it really acts as a hook
to style variable HTML a certain way. This allows a developer to wrap
any markup in documentation and it will all be styled consistently with
the panel.