Expandable box component
And fixes to pattern library with all the panel row stuff being
taken out.

Expandable box is a component to take care of the overview page
quicklooks where you can click on one to expand it. I also changed
card the card and panel features to more specific to actual cloud.gov.