Refactor usage and limits
By removing panel code and making the stat component much more
specific. Rather then making the stat component able to adapt
to many different sizes, it sets all the sizes of things. This
makes it much easier to build and I'm not sure the adaptability
will be required for this component anyway.