Rewrote notifications based on design feedback
To have a outline button with transparent background of the color
of the notification.

To get this to work, I did some work to remove the massive amount
of specificity in the buttons. This was mainly done by removing the
dependency on the `usa-button` classes which have too much specificty.

I also ensured the styleguide uses unmin css because even in production
it seems more useful.

This is a breaking change.